LPS-100795 Move to root folder
