+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/table/DropDown.es.js
@@ -16,7 +16,7 @@ import ClayDropDown, {Align} from '@clayui/drop-down';
 import ClayIcon from '@clayui/icon';
 import React, {useState} from 'react';
 import DropDownAction from './DropDownAction.es';
-import Button from '../../button/Button.es';
+import Button from '../button/Button.es';
 
 const {ItemList} = ClayDropDown;
