+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/table/EmptyState.es.js
@@ -13,7 +13,7 @@
  */
 
 import React from 'react';
-import lang from '../../../utils/lang.es';
+import lang from '../../utils/lang.es';
 
 const EmptyState = ({emptyState, keywords = ''}) => {
 	const defaultEmpty = {
