+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/table/Table.es.js
